Update code list 'create/update/delete' logic in regard with manifest.

+++ src/main/java/org/oagi/srt/gateway/http/api/code_list_management/data/DeleteCodeListRequest.java
@@ -2,11 +2,12 @@ package org.oagi.srt.gateway.http.api.code_list_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 
 @Data
 public class DeleteCodeListRequest {
 
-    private List<Long> codeListIds = Collections.emptyList();
+    private List<BigInteger> codeListIds = Collections.emptyList();
 }
